Refactor library into concept of Words ahead of rich text support
__WARNING:__ Still a WIP. See `// DEBUG` comments throughout.

Instead of dealing with text as a string, introduce the concept of Words
where a Word is either a visible, written word, or whitespace, and have
the option to infer whitespace when an array of Words is given instead
of a string.

A Word then has its own formatting options aside from the "base" formatting
set in the `CanvasTextConfig` object. If a Word doesn't have `format` options,
it will simply use the "default" format in the 2D context, which may have been
modified based on the `CanvasTextConfig` object provided to `drawText()`.

To preserve the existing `drawText()` and `splitText()` APIs, `splitText()`
splits the text into an array of Words and then calls the new `splitWords()`
function that does all the work.

For now, providing an array of Words to `drawText()` throws an error since
this isn't supported yet. But everything else workds as before when providing
a string: All supported alignments, including justification.

Other fixes/changes:
- Breaks cannot occur within a word anymore, even when the box width gets
  too narrow to fit anything. At this point, a minimum of one word is rendered
  per line, on top of any hard line breaks specified with newline characters.
- The context's state is saved and restored in `drawText()`, meaning that
  after the call, the context's font, etc, styles/settings are back to what
  they were prior to calling it instead of being set to what was specified
  in the `CanvasTextConfig` object.