BREAKING: Move text placement into splitWords()
__WARNING:__ See remaining `// DEBUG` comments.

In order to make it possible to use a library other than canvas-txt to
actually draw the text, yet still leverage this library's text wrapping
and rich text features, `splitWords()` now does all the positioning
that `drawText()` use to do, which means that `splitText()`, which calls
`splitWords()` internally, must now be given many more parameters than
it used to.

Since `splitText()` was already exported as a utility from this library,
this must be a breaking change (major version update).

So now, `splitWords()` returns the positioned words along with the
required `textAlign` and `textBaseline` to set on the 2D context
in order to render the text what is now a very simple loop.

This makes it possible to call `splitWords()` directly and then use
the output to render each word using another library such as Konva
where each word could then be dragged or hit-tested, or do it
without a library altogether.

Plain text (single text format) works at this point, but rich text
still does not render properly due to misalignment issues.