Fix rich text rendering
__WARNING:__ See `// DEBUG` comments for what still needs doing/fixing.

Works quite well now, especially when NOT using different font sizes
for different words. If all words use the same size, just alter their
style, weight, or variant, it renders nicely.

Updated the demo app to always set "ipsum" to italic, and "consectetur"
to bold, if either is found in the sample text to render.