Work with TextMetrics-like objects
Native Canvas `TextMetrics` objects fail to serialize (cause an exception)
when given as data to the `postMessage()` of a Web Worker thread (or from
the main thread).

Therefore, work with objects having a similar interface but not being
actual `TextMetrics` instances.

Also provide two JSON stringify helpers: `specToJson()` and `wordsToJson()`
This should help consumers to deal with native `TextMetrics` objects.